SUS C5 Version 1.0
Version 1.0 - Replaced Display Name Function in State Objects with Send Response Function that checks if its name matches the name of the state currently selected; if not, it displays as red with name, if so, it displays as green with name with 'Correct' in question label. Nothing occurs if no question is current (including if the question has been answered). Every new question clears all formatted states.